% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completion.R
\name{arg_completion}
\alias{arg_completion}
\title{complete a function argument}
\usage{
arg_completion(workspace, token, call)
}
\arguments{
\item{workspace}{a \link{Workspace} object}

\item{token}{a character, the start of the argument to identify}

\item{call}{a call object, the function for which \code{token} may be an argument}
}
\value{
a list of candidates
}
\description{
complete a function argument
}
