% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{call_hierarchy_incoming_calls}
\alias{call_hierarchy_incoming_calls}
\title{\code{callHierarchy/incomingCalls} request handler}
\usage{
call_hierarchy_incoming_calls(self, id, params)
}
\description{
Handler to the \code{callHierarchy/incomingCalls} \link{Request}.
}
\keyword{internal}
