% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{call_hierarchy_outgoing_calls}
\alias{call_hierarchy_outgoing_calls}
\title{\code{callHierarchy/outgoingCalls} request handler}
\usage{
call_hierarchy_outgoing_calls(self, id, params)
}
\description{
Handler to the \code{callHierarchy/outgoingCalls} \link{Request}.
}
\keyword{internal}
