% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-general.R
\name{cancel_request}
\alias{cancel_request}
\title{\code{cancel} request notification handler}
\usage{
cancel_request(self, params)
}
\description{
Handler to the \code{cancelRequest} \link{Notification}.
}
\keyword{internal}
