% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_scope}
\alias{check_scope}
\title{Check if a token is in a R code block in an Rmarkdown file}
\usage{
check_scope(uri, document, point)
}
\description{
In an RMarkdown document, tokens can be either inside an R code block or
in the text. This function will return \code{FALSE} if the token is in the text
and \code{TRUE} if it is in a code block. For any other files, it always returns \code{TRUE}.
}
\keyword{internal}
