% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{code_action_params}
\alias{code_action_params}
\title{Parameters for code action requests}
\usage{
code_action_params(uri, range, context = NULL)
}
\arguments{
\item{uri}{a character, the path to a file as defined by \href{https://tools.ietf.org/html/rfc3986}{RFC 3986}}

\item{range}{a \link{range} object}

\item{context}{a named list}
}
\description{
Parameters for code action requests
}
\keyword{internal}
