% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{code_point_from_unit}
\alias{code_point_from_unit}
\title{Determinal code points given code units}
\usage{
code_point_from_unit(line, units)
}
\arguments{
\item{line}{a character of text}

\item{units}{0-indexed code points}
}
\description{
Determinal code points given code units
}
\keyword{internal}
