% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{code_point_to_unit}
\alias{code_point_to_unit}
\title{Determine code units given code points}
\usage{
code_point_to_unit(line, pts)
}
\arguments{
\item{line}{a character of text}

\item{units}{0-indexed code units}
}
\description{
Determine code units given code points
}
\keyword{internal}
