% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color_presentation_reply}
\alias{color_presentation_reply}
\title{The response to a textDocument/colorPresentation Request}
\usage{
color_presentation_reply(id, uri, workspace, document, color)
}
\description{
The response to a textDocument/colorPresentation Request
}
\keyword{internal}
