% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition.R
\name{definition_reply}
\alias{definition_reply}
\title{Get the location of a specified function definition}
\usage{
definition_reply(id, uri, workspace, document, position)
}
\description{
If the function is not found in a file but is found in a loaded package,
writes the function definition to a temporary file and returns that
as the location.
}
\keyword{internal}
