% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_function}
\alias{detect_function}
\title{check if a character vector looks like a function}
\usage{
detect_function(text)
}
\arguments{
\item{text}{a character vector}
}
\description{
check if a character vector looks like a function
}
\keyword{internal}
