% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{detect_hover}
\alias{detect_hover}
\title{detect the token at the current position}
\usage{
detect_hover(document, position)
}
\arguments{
\item{document}{a character, the contents of the document}

\item{position}{a \link{position} object}
}
\description{
detect the token at the current position
}
