% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{detect_token}
\alias{detect_token}
\title{detect if position contains a valid token}
\usage{
detect_token(document, position)
}
\arguments{
\item{document}{a character, the contents of the document}

\item{position}{a \link{position} object}
}
\description{
detect if position contains a valid token
}
