% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnose_file}
\alias{diagnose_file}
\title{run diagnostic on a file}
\usage{
diagnose_file(path)
}
\arguments{
\item{path}{a character, the path to a file}
}
\description{
Lint and diagnose problems in a file.
}
