% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{did_change_configuration_params}
\alias{did_change_configuration_params}
\title{Parameters for workspace/didChangeConfiguration notifications}
\usage{
did_change_configuration_params(settings)
}
\arguments{
\item{settings}{a named list}
}
\description{
Parameters for workspace/didChangeConfiguration notifications
}
\keyword{internal}
