% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{did_close_text_document_params}
\alias{did_close_text_document_params}
\title{parameters for didClose notifications}
\usage{
did_close_text_document_params(uri)
}
\arguments{
\item{uri}{a character, the path to a file as defined by \href{https://tools.ietf.org/html/rfc3986}{RFC 3986}}
}
\description{
parameters for didClose notifications
}
