% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{did_save_text_document_params}
\alias{did_save_text_document_params}
\title{Parameters for didSave notifications}
\usage{
did_save_text_document_params(uri, text)
}
\arguments{
\item{uri}{a character, the path to a file as defined by \href{https://tools.ietf.org/html/rfc3986}{RFC 3986}}

\item{text}{a character}
}
\description{
Parameters for didSave notifications
}
\keyword{internal}
