% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document_backward_search}
\alias{document_backward_search}
\title{search backwards in a document for a specific character}
\usage{
document_backward_search(document, line, character, char,
  skip_empty_line = TRUE)
}
\arguments{
\item{document}{a character, the contents of the document}

\item{line}{a numeric, the line number}

\item{character}{a numeric, the column number of the character}

\item{char}{a single character}

\item{skip_empty_line}{a logical}
}
\value{
a tuple of positive integers, the line and column position of the
character if found, otherwise (-1, -1)
}
\description{
search backwards in a document for a specific character
}
