% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_action.R
\name{document_code_action_reply}
\alias{document_code_action_reply}
\title{The response to a textDocument/codeAction Request}
\usage{
document_code_action_reply(
  id,
  uri,
  workspace,
  document,
  range,
  context,
  client_capabilities = NULL
)
}
\description{
The response to a textDocument/codeAction Request
}
\keyword{internal}
