% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{document_color_reply}
\alias{document_color_reply}
\title{The response to a textDocument/documentColor Request}
\usage{
document_color_reply(id, uri, workspace, document)
}
\description{
The response to a textDocument/documentColor Request
}
\keyword{internal}
