% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folding.R
\name{document_folding_range_reply}
\alias{document_folding_range_reply}
\title{Get all the folding ranges in the document}
\usage{
document_folding_range_reply(id, uri, workspace, document)
}
\description{
Get all the folding ranges in the document
}
\keyword{internal}
