% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{document_formatting_params}
\alias{document_formatting_params}
\title{parameters for document formatting requests}
\usage{
document_formatting_params(uri, options)
}
\arguments{
\item{uri}{a character, the path to a file as defined by \href{https://tools.ietf.org/html/rfc3986}{RFC 3986}}

\item{options}{a named list}
}
\description{
parameters for document formatting requests
}
