% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{document_highlight_reply}
\alias{document_highlight_reply}
\title{The response to a textDocument/documentHighlight Request}
\usage{
document_highlight_reply(id, uri, workspace, document, point)
}
\description{
The response to a textDocument/documentHighlight Request
}
\keyword{internal}
