% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document_line}
\alias{document_line}
\title{get the contents of a line}
\usage{
document_line(document, lineno)
}
\arguments{
\item{document}{a character, the contents of the document}

\item{lineno}{a numeric, the line number}
}
\value{
a character
}
\description{
If the line number is higher than the number of lines in the document,
an empty character is returned.
}
