% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{document_link_params}
\alias{document_link_params}
\title{Parameters for document link requests}
\usage{
document_link_params(uri)
}
\arguments{
\item{uri}{a character, the path to a file as defined by \href{https://tools.ietf.org/html/rfc3986}{RFC 3986}}
}
\description{
Parameters for document link requests
}
\keyword{internal}
