% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{document_link_reply}
\alias{document_link_reply}
\title{The response to a textDocument/documentLink Request}
\usage{
document_link_reply(id, uri, workspace, document, rootPath)
}
\arguments{
\item{rootPath}{Path of workspace folder}
}
\description{
The response to a textDocument/documentLink Request
}
\keyword{internal}
