% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{document_link_resolve}
\alias{document_link_resolve}
\title{\code{documentLink/resolve} request handler}
\usage{
document_link_resolve(self, id, params)
}
\description{
Handler to the \code{documentLink/resolve} \link{Request}.
}
\keyword{internal}
