% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{document_on_type_formatting_params}
\alias{document_on_type_formatting_params}
\title{parameters for document on type formatting requests}
\usage{
document_on_type_formatting_params(uri, position, character, options)
}
\arguments{
\item{uri}{a character, the path to a file as defined by \href{https://tools.ietf.org/html/rfc3986}{RFC 3986}}

\item{position}{a \link{position} object}

\item{character}{a single character}

\item{options}{a named list}
}
\description{
parameters for document on type formatting requests
}
