% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{document_range_formatting_params}
\alias{document_range_formatting_params}
\title{Parameters for document range formatting requests}
\usage{
document_range_formatting_params(uri, range, options)
}
\arguments{
\item{uri}{a character, the path to a file as defined by \href{https://tools.ietf.org/html/rfc3986}{RFC 3986}}

\item{range}{a \link{range} object}

\item{options}{a named list}
}
\description{
Parameters for document range formatting requests
}
\keyword{internal}
