% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{document_symbol_reply}
\alias{document_symbol_reply}
\title{Get all the symbols in the document}
\usage{
document_symbol_reply(id, uri, workspace, document, capabilities)
}
\description{
Get all the symbols in the document
}
\keyword{internal}
