% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{enclosed_by_quotes}
\alias{enclosed_by_quotes}
\title{check if a position is inside quotes}
\usage{
enclosed_by_quotes(document, pos)
}
\description{
check if a position is inside quotes
}
\keyword{internal}
