% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_blocks}
\alias{extract_blocks}
\title{Extract the R code blocks of a Rmarkdown file}
\usage{
extract_blocks(content)
}
\description{
Extract the R code blocks of a Rmarkdown file
}
\keyword{internal}
