% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{find_config}
\alias{find_config}
\title{find the lintr config file}
\usage{
find_config(filename)
}
\description{
This is a copy of \code{lintr:::find_config} since CRAN doesn't like :::
}
\keyword{internal}
