% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_package}
\alias{find_package}
\title{Find the root package folder}
\usage{
find_package(path = getwd())
}
\description{
This function searches backwards in the folder structure until it finds
a DESCRIPTION file or it reaches the top-level directory.
}
\keyword{internal}
