% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{formatting_reply}
\alias{formatting_reply}
\title{format a document}
\usage{
formatting_reply(id, uri, document, options)
}
\arguments{
\item{id}{a numeric, the id of the process that started the server}

\item{uri}{a character, the path to a file as defined by \href{https://tools.ietf.org/html/rfc3986}{RFC 3986}}

\item{document}{a character, the contents of the document}

\item{options}{a named list of options, with a \code{tabSize} parameter}
}
\description{
format a document
}
