% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\docType{data}
\name{GlobalNameSpace}
\alias{GlobalNameSpace}
\title{A class for storing global environment information}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
GlobalNameSpace
}
\description{
A class for storing global environment information
}
\keyword{datasets}
