% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_directory}
\alias{is_directory}
\title{check if a filename is a directory}
\usage{
is_directory(filename)
}
\arguments{
\item{filename}{a character}
}
\description{
check if a filename is a directory
}
\keyword{internal}
