% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_package}
\alias{is_package}
\title{check if an URI is a package folder}
\usage{
is_package(rootUri)
}
\arguments{
\item{rootUri}{a character representing a URI}
}
\description{
check if an URI is a package folder
}
\keyword{internal}
