% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{is_rmarkdown}
\alias{is_rmarkdown}
\title{check if a file is an RMarkdown file}
\usage{
is_rmarkdown(uri)
}
\arguments{
\item{uri}{a character, the path to a file as defined by \href{https://tools.ietf.org/html/rfc3986}{RFC 3986}}
}
\value{
a logical
}
\description{
check if a file is an RMarkdown file
}
