% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageserver.R
\docType{data}
\name{LanguageServer}
\alias{LanguageServer}
\title{the language server}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LanguageServer
}
\description{
Describe the language server and how it interacts with clients.
}
\keyword{datasets}
