% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{leisurize}
\alias{leisurize}
\title{delay a function execution}
\usage{
leisurize(fun, t = 1)
}
\arguments{
\item{fun}{the function to execute}

\item{t}{an integer, the threshold in seconds}
}
\description{
Execute a function if the last execution time is older than a specified
value.
}
\keyword{internal}
