% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{location}
\alias{location}
\title{a location inside a resource}
\usage{
location(uri, range)
}
\arguments{
\item{uri}{a character, the path to a file}

\item{range}{a \link{range}}
}
\description{
a location inside a resource
}
