% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{logger}
\alias{logger}
\title{create the logger}
\format{An object of class \code{Logger} (inherits from \code{R6}) of length 5.}
\usage{
logger
}
\description{
create the logger
}
\keyword{internal}
