% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\docType{data}
\name{Namespace}
\alias{Namespace}
\title{A data structure for package namespaces}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Namespace
}
\description{
A \code{Namespace} is initialized with a package name and builds the list of
objects defined in the package namespace.
}
\keyword{internal}
