% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\docType{data}
\name{Notification}
\alias{Notification}
\title{Notification Message class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Notification
}
\description{
Notification Message class
}
\keyword{internal}
