% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-general.R
\name{on_exit}
\alias{on_exit}
\title{exit notification handler}
\usage{
on_exit(self, params)
}
\arguments{
\item{params}{unused here}
}
\description{
Hanlder to the \href{https://microsoft.github.io/language-server-protocol/specification#exit}{exit} \link{Notification}
}
\keyword{internal}
