% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-general.R
\name{on_initialize}
\alias{on_initialize}
\title{\code{initialize} handler}
\usage{
on_initialize(self, id, params)
}
\description{
Handler to the \code{initialize} \link{Request}.
}
\keyword{internal}
