% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-general.R
\name{on_initialized}
\alias{on_initialized}
\title{\code{initialized} handler}
\usage{
on_initialized(self, params)
}
\description{
Handler to the \code{initialized} \link{Notification}.
}
\keyword{internal}
