% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{on_type_formatting_reply}
\alias{on_type_formatting_reply}
\title{Format on type}
\usage{
on_type_formatting_reply(id, uri, document, point, ch, options)
}
\description{
Format on type
}
\keyword{internal}
