% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\docType{data}
\name{PackageNamespace}
\alias{PackageNamespace}
\title{A class for storing package information}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
PackageNamespace
}
\description{
A class for storing package information
}
\keyword{internal}
