% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{parse_document}
\alias{parse_document}
\title{Parse a document}
\usage{
parse_document(uri, content, resolve = FALSE)
}
\description{
Build the list of called packages, functions, variables, formals and
signatures in the document in order to add them to the current \link{Workspace}.
}
\keyword{internal}
