% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_from_uri}
\alias{path_from_uri}
\alias{path_to_uri}
\title{paths and uris}
\usage{
path_from_uri(uri)

path_to_uri(path)
}
\arguments{
\item{uri}{a character, the path to a file in URI format}

\item{path}{a character, the path to a file}
}
\description{
paths and uris
}
