% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_is_detached}
\alias{process_is_detached}
\title{check if the current process becomes an orphan}
\usage{
process_is_detached()
}
\description{
check if the current process becomes an orphan
}
\keyword{internal}
