% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{purl}
\alias{purl}
\title{Strip out all the non R blocks in a R markdown file}
\usage{
purl(content)
}
\arguments{
\item{content}{a character vector}
}
\description{
Strip out all the non R blocks in a R markdown file
}
\keyword{internal}
