% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{range}
\alias{range}
\title{A range in a text document}
\usage{
range(start, end)
}
\arguments{
\item{start}{a \link{position}}

\item{end}{a \link{position}}
}
\description{
A range in a text document expressed as start and end \link{position}s.
A range is comparable to a selection in an editor.
}
\keyword{internal}
