% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{range_formatting_reply}
\alias{range_formatting_reply}
\title{Format a part of a document}
\usage{
range_formatting_reply(id, uri, document, range, options)
}
\description{
Format a part of a document
}
\keyword{internal}
