% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{reference_params}
\alias{reference_params}
\title{parameters for reference requests}
\usage{
reference_params(uri, position, context = NULL)
}
\arguments{
\item{uri}{a character, the path to a file as defined by \href{https://tools.ietf.org/html/rfc3986}{RFC 3986}}

\item{position}{a \link{position} object}

\item{context}{a named list}
}
\description{
parameters for reference requests
}
