% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{rename_params}
\alias{rename_params}
\title{parameters for rename requests}
\usage{
rename_params(uri, position, new_name)
}
\arguments{
\item{uri}{a character, the path to a file as defined by \href{https://tools.ietf.org/html/rfc3986}{RFC 3986}}

\item{position}{a \link{position} object}

\item{new_name}{a character}
}
\description{
parameters for rename requests
}
