% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\docType{data}
\name{Request}
\alias{Request}
\title{Request Message class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Request
}
\description{
Describe a request between the client and the server.
}
\keyword{datasets}
