% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\docType{data}
\name{Response}
\alias{Response}
\title{Response Message class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Response
}
\description{
Message sent as the result of a \link{Request}
}
\keyword{internal}
