% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\docType{data}
\name{ResponseErrorMessage}
\alias{ResponseErrorMessage}
\title{Response Error Message class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ResponseErrorMessage
}
\description{
Message sent as the result of a \link{Request} in case of an error.
}
\keyword{internal}
