% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageserver.R
\name{run}
\alias{run}
\title{Run the R language server}
\usage{
run(debug = FALSE, host = "localhost", port = NULL)
}
\arguments{
\item{debug}{set \code{TRUE} to show debug information in stderr}

\item{host}{the hostname used to create the tcp server, not used when \code{port} is \code{NULL}}

\item{port}{the port used to create the tcp server. If \code{NULL}, use stdio instead.}
}
\description{
Run the R language server
}
