% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_seq}
\alias{safe_seq}
\title{Safer version of \code{seq} which returns empty vector if b < a}
\usage{
safe_seq(a, b)
}
\description{
Safer version of \code{seq} which returns empty vector if b < a
}
\keyword{internal}
