% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize_names}
\alias{sanitize_names}
\title{sanitize package objects names}
\usage{
sanitize_names(objects)
}
\arguments{
\item{objects}{a character vector}
}
\description{
Remove unwanted objects, \emph{e.g.} objects whose name starts with \code{_} or \code{.}.
}
\keyword{internal}
