% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{selection_range_reply}
\alias{selection_range_reply}
\title{The response to a textDocument/selectionRange Request}
\usage{
selection_range_reply(id, uri, workspace, document, points)
}
\description{
The response to a textDocument/selectionRange Request
}
\keyword{internal}
