% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_safe}
\alias{seq_safe}
\title{Safer version of \code{seq} which returns empty vector if b < a}
\usage{
seq_safe(a, b)
}
\description{
Safer version of \code{seq} which returns empty vector if b < a
}
\keyword{internal}
