% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capabilities.R
\name{ServerCapabilities}
\alias{ServerCapabilities}
\title{Server capabilities}
\description{
List all supported capabilities of the language server, as defined in the
\href{https://microsoft.github.io/language-server-protocol/specification}{Language Server Protocol specification}.
}
\details{
Currently the capabilities are:
\itemize{
\item textDocumentSync,
\item hoverProvider,
\item completionProvider,
\item signatureHelpProvider,
\item documentFormattingProvider,
\item documentRangeFormattingProvider
}
}
