% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{SessionPool}
\alias{SessionPool}
\title{Session Pool with Many R Sessions}
\description{
Session Pool with Many R Sessions
}
\examples{
\dontrun{
# create 3 cached r sessions in the pool
pool_size <- 3
pool_name <- "common"
pool <- SessionPool$new(pool_size, pool_name)

# check idle_size before acquiring session
n <- pool$get_idle_size()

# if there are idle sessions, acquire session
session <- pool$acquire()

# check session is not null and then use session
is.null(session)

# use session
# please read R6 Class `Session` documentation

}
}
