% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{signature_reply}
\alias{signature_reply}
\title{the response to a textDocument/signatureHelp Request}
\usage{
signature_reply(id, uri, workspace, document, position)
}
\arguments{
\item{id}{a numeric, the id of the process that started the server}

\item{uri}{a character, the path to a file as defined by \href{https://tools.ietf.org/html/rfc3986}{RFC 3986}}

\item{workspace}{a \link{Workspace} object, the current workspace}

\item{document}{a character, the contents of the document}

\item{position}{a \link{position} object}
}
\value{
a \link{Response} object
}
\description{
If the symbol at the current position is a function, return its arguments
(as with \code{\link[base:args]{base::args()}}).
}
