% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{style_text}
\alias{style_text}
\title{edit code style}
\usage{
style_text(text, options)
}
\arguments{
\item{text}{a list of text}

\item{options}{a named list of options, with a \code{tabSize} parameter}
}
\description{
This functions formats a list of text using \code{\link[styler:style_text]{styler::style_text()}} with the
\code{\link[styler:tidyverse_style]{styler::tidyverse_style()}} style.
}
