% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{stylize}
\alias{stylize}
\title{edit code style}
\usage{
stylize(document, options)
}
\arguments{
\item{document}{a character, the contents of the document}

\item{options}{a named list of options, with a \code{tabSize} parameter}
}
\description{
This functions formats a document using \code{\link[styler:style_text]{styler::style_text()}} with the
\code{\link[styler:tidyverse_style]{styler::tidyverse_style()}} style.
}
