% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{symbol_information}
\alias{symbol_information}
\title{non-hierachical symbol information}
\usage{
symbol_information(name, kind, location)
}
\arguments{
\item{name}{a character}

\item{kind}{an integer}

\item{location}{a \link{location}}
}
\description{
non-hierachical symbol information
}
