% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_code_action}
\alias{text_document_code_action}
\title{\code{textDocument/codeAction} request handler}
\usage{
text_document_code_action(self, id, params)
}
\description{
Handler to the \code{textDocument/codeAction} \link{Request}.
}
\keyword{internal}
