% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_code_lens}
\alias{text_document_code_lens}
\title{\code{textDocument/codeLens} request handler}
\usage{
text_document_code_lens(self, id, params)
}
\description{
Handler to the \code{textDocument/codeLens} \link{Request}.
}
\keyword{internal}
