% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_color_presentation}
\alias{text_document_color_presentation}
\title{\code{textDocument/colorPresentation} request handler}
\usage{
text_document_color_presentation(self, id, params)
}
\description{
Handler to the \code{textDocument/colorPresentation} \link{Request}.
}
\keyword{internal}
