% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_completion}
\alias{text_document_completion}
\title{\code{textDocument/completion} request handler}
\usage{
text_document_completion(self, id, params)
}
\description{
Handler to the \code{textDocument/completion} \link{Request}.
}
\keyword{internal}
