% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-textsync.R
\name{text_document_did_close}
\alias{text_document_did_close}
\title{\code{textDocument/didClose} notification handler}
\usage{
text_document_did_close(self, params)
}
\description{
Handler to the \code{textDocument/didClose} \link{Notification}.
}
\keyword{internal}
