% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-textsync.R
\name{text_document_did_open}
\alias{text_document_did_open}
\title{\code{textDocument/didOpen} notification handler}
\usage{
text_document_did_open(self, params)
}
\description{
Handler to the \code{textDocument/didOpen} \link{Notification}.
}
\keyword{internal}
