% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-textsync.R
\name{text_document_did_save}
\alias{text_document_did_save}
\title{\code{textDocument/didSave} notification handler}
\usage{
text_document_did_save(self, params)
}
\description{
Handler to the \code{textDocument/didSave} \link{Notification}.
}
\keyword{internal}
