% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_document_link}
\alias{text_document_document_link}
\title{\code{textDocument/documentLink} request handler}
\usage{
text_document_document_link(self, id, params)
}
\description{
Handler to the \code{textDocument/documentLink} \link{Request}.
}
\keyword{internal}
