% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_document_symbol}
\alias{text_document_document_symbol}
\title{\code{textDocument/documentSymbol} request handler}
\usage{
text_document_document_symbol(self, id, params)
}
\description{
Handler to the \code{textDocument/documentSymbol} \link{Request}.
}
\keyword{internal}
