% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_folding_range}
\alias{text_document_folding_range}
\title{\code{textDocument/foldingRange} request handler}
\usage{
text_document_folding_range(self, id, params)
}
\description{
Handler to the \code{textDocument/foldingRange} \link{Request}.
}
\keyword{internal}
