% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_formatting}
\alias{text_document_formatting}
\title{\code{textDocument/formatting} request handler}
\usage{
text_document_formatting(self, id, params)
}
\description{
Handler to the \code{textDocument/formatting} \link{Request}.
}
\keyword{internal}
