% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_hover}
\alias{text_document_hover}
\title{\code{textDocument/hover} request handler}
\usage{
text_document_hover(self, id, params)
}
\description{
Handler to the \code{textDocument/hover} \link{Request}.
}
\keyword{internal}
