% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_implementation}
\alias{text_document_implementation}
\title{\code{textDocument/implementation} request handler}
\usage{
text_document_implementation(self, id, params)
}
\description{
Handler to the \code{textDocument/implementation} \link{Request}.
}
\keyword{internal}
