% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{text_document_position_params}
\alias{text_document_position_params}
\title{A text document and a position inside that document}
\usage{
text_document_position_params(uri, position)
}
\arguments{
\item{uri}{a character, the path to a file as defined by \href{https://tools.ietf.org/html/rfc3986}{RFC 3986}}

\item{position}{a \link{position} object}
}
\description{
A text document and a position inside that document
}
\keyword{internal}
