% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_prepare_call_hierarchy}
\alias{text_document_prepare_call_hierarchy}
\title{\code{textDocument/prepareCallHierarchy} request handler}
\usage{
text_document_prepare_call_hierarchy(self, id, params)
}
\description{
Handler to the \code{textDocument/prepareCallHierarchy} \link{Request}.
}
\keyword{internal}
