% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_prepare_rename}
\alias{text_document_prepare_rename}
\title{\code{textDocument/prepareRename} request handler}
\usage{
text_document_prepare_rename(self, id, params)
}
\description{
Handler to the \code{textDocument/prepareRename} \link{Request}.
}
\keyword{internal}
