% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_references}
\alias{text_document_references}
\title{\code{textDocument/references} request handler}
\usage{
text_document_references(self, id, params)
}
\description{
Handler to the \code{textDocument/references} \link{Request}.
}
\keyword{internal}
