% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_signature_help}
\alias{text_document_signature_help}
\title{\code{textDocument/signatureHelp} request handler}
\usage{
text_document_signature_help(self, id, params)
}
\description{
Handler to the \code{textDocument/signatureHelp} \link{Request}.
}
\keyword{internal}
