% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_type_definition}
\alias{text_document_type_definition}
\title{\code{textDocument/typeDefinition} request handler}
\usage{
text_document_type_definition(self, id, params)
}
\description{
Handler to the \code{textDocument/typeDefinition} \link{Request}.
}
\keyword{internal}
