% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-textsync.R
\name{text_document_will_save_wait_until}
\alias{text_document_will_save_wait_until}
\title{\code{textDocument/willSaveWaitUntil} notification handler}
\usage{
text_document_will_save_wait_until(self, id, params)
}
\description{
Handler to the \code{textDocument/willSaveWaitUntil} \link{Request}.
}
\keyword{internal}
