% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{throttle}
\alias{throttle}
\title{throttle a function execution}
\usage{
throttle(fun, t = 1)
}
\arguments{
\item{fun}{the function to execute}

\item{t}{an integer, the threshold in seconds}
}
\description{
Execute a function if the last execution time is older than a specified
value.
}
\keyword{internal}
