% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{token_starts_with_s3}
\alias{token_starts_with_s3}
\title{Does the current token start with a valid S3 generic}
\usage{
token_starts_with_s3(token)
}
\arguments{
\item{token}{A character value to consider}
}
\description{
Does the current token start with a valid S3 generic
}
\examples{
token_starts_with_s3("as.character.fac")
# TRUE

token_starts_with_s3("as.chara")
# FALSE

}
\keyword{internal}
