% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utf8_to_utf16_code_point}
\alias{utf8_to_utf16_code_point}
\title{determinal utf16 code points given utf8 code points}
\usage{
utf8_to_utf16_code_point(line, pts)
}
\arguments{
\item{line}{a character of text}

\item{pts}{0-indexed code points}
}
\description{
determinal utf16 code points given utf8 code points
}
\keyword{internal}
