% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{will_save_text_document_params}
\alias{will_save_text_document_params}
\title{Parameters for willSave notifications}
\usage{
will_save_text_document_params(uri, reason)
}
\arguments{
\item{uri}{a character, the path to a file as defined by \href{https://tools.ietf.org/html/rfc3986}{RFC 3986}}

\item{reason}{an integer, see TextDocumentSaveReason}
}
\description{
Parameters for willSave notifications
}
\keyword{internal}
