% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-workspace.R
\name{workspace_did_change_configuration}
\alias{workspace_did_change_configuration}
\title{\code{workspace/didChangeConfiguration} notification handler}
\usage{
workspace_did_change_configuration(self, params)
}
\description{
Handler to the \code{workspace/didChangeConfiguration} \link{Notification}
}
\keyword{internal}
