% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-workspace.R
\name{workspace_did_change_watched_files}
\alias{workspace_did_change_watched_files}
\title{workspace/didChangeWatchedFiles notification handler}
\usage{
workspace_did_change_watched_files(self, params)
}
\arguments{
\item{self}{a \link{LanguageServer} object}

\item{params}{a did_change_watched_files_params object}
}
\description{
Handler to the \href{https://microsoft.github.io/language-server-protocol/}{workspace/didChangeWatchedFiles} \link{Notification}.
}
\details{
Not implemented yet.
}
\keyword{internal}
