% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-workspace.R
\name{workspace_did_change_workspace_folder_params}
\alias{workspace_did_change_workspace_folder_params}
\title{\code{workspace/didChangeWorkspaceFolders} notification handler}
\usage{
workspace_did_change_workspace_folder_params(self, params)
}
\description{
Handler to the \code{workspace/didChangeWorkspaceFolders} \link{Notification}.
}
\keyword{internal}
