% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-workspace.R
\name{workspace_execute_command}
\alias{workspace_execute_command}
\title{\code{workspace/executeCommand} request handler}
\usage{
workspace_execute_command(self, id, params)
}
\description{
Handler to the \code{workspace/executeCommand} \link{Request}.
}
\keyword{internal}
