% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-workspace.R
\name{workspace_symbol}
\alias{workspace_symbol}
\title{\code{workspace/symbol} request handler}
\usage{
workspace_symbol(self, id, params)
}
\description{
Handler to the \code{workspace/symbol} \link{Request}.
}
\keyword{internal}
