% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{workspace_sync}
\alias{workspace_sync}
\title{Determine workspace information for a given file}
\usage{
workspace_sync(uri, run_lintr = TRUE, parse_file = FALSE,
  temp_file = NULL)
}
\arguments{
\item{uri}{the file path}

\item{run_lintr}{set \code{FALSE} to disable lintr diagnostics}

\item{parse_file}{set \code{FALSE} to disable lintr diagnostics}

\item{temp_file}{the actual file to lint}
}
\description{
internal use only
}
