% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completion.R
\name{arg_completion}
\alias{arg_completion}
\title{Complete a function argument}
\usage{
arg_completion(
  uri,
  workspace,
  point,
  token,
  funct,
  package = NULL,
  exported_only = TRUE
)
}
\description{
Complete a function argument
}
\keyword{internal}
