% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-general.R
\name{cancel_request}
\alias{cancel_request}
\title{cancel request notification handler}
\usage{
cancel_request(self, params)
}
\arguments{
\item{params}{unused here}
}
\description{
Handler to the \href{https://microsoft.github.io/language-server-protocol/specification#cancelRequest}{cancelRequest} \link{Notificatio}
}
\keyword{internal}
