% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{code_lens_params}
\alias{code_lens_params}
\title{Parameters for code lens requests}
\usage{
code_lens_params(uri)
}
\arguments{
\item{uri}{a character, the path to a file as defined by \href{https://tools.ietf.org/html/rfc3986}{RFC 3986}}
}
\description{
Parameters for code lens requests
}
\keyword{internal}
