% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{code_lens_resolve}
\alias{code_lens_resolve}
\title{\code{codeLens/resolve} request handler}
\usage{
code_lens_resolve(self, id, params)
}
\description{
Handler to the \code{codeLens/resolve} \link{Request}.
}
\keyword{internal}
