% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{code_point_to_unit}
\alias{code_point_to_unit}
\title{determinal code units given code points}
\usage{
code_point_to_unit(line, pts)
}
\arguments{
\item{line}{a character of text}

\item{pts}{0-indexed code points}
}
\description{
determinal code units given code points
}
\keyword{internal}
