% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{completion_item_resolve}
\alias{completion_item_resolve}
\title{\code{completionItem/resolve} request handler}
\usage{
completion_item_resolve(self, id, params)
}
\description{
Handler to the \code{completionItem/resolve} \link{Request}.
}
\keyword{internal}
