% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completion.R
\name{completion_reply}
\alias{completion_reply}
\title{The response to a textDocument/completion request}
\usage{
completion_reply(id, uri, workspace, document, position)
}
\description{
The response to a textDocument/completion request
}
\keyword{internal}
