% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completion.R
\name{completion_reply}
\alias{completion_reply}
\title{the response to a textDocument/completion request}
\usage{
completion_reply(id, uri, workspace, document, position)
}
\value{
a \link{Response} object

a \link{Response} object
}
\description{
the response to a textDocument/completion request
}
