% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\docType{data}
\name{DefinitionCache}
\alias{DefinitionCache}
\title{A data structure to hold function definition locations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DefinitionCache
}
\description{
The key reason for using this rather than a \code{list} is that this also cleans up
when functions are removed from files.
}
\keyword{internal}
