% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{detect_closure}
\alias{detect_closure}
\title{detect if the current position is inside a closure}
\usage{
detect_closure(document, position)
}
\arguments{
\item{document}{a character, the contents of the document}

\item{position}{a \link{position} object}
}
\description{
detect if the current position is inside a closure
}
