% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\alias{diagnostic_range}
\alias{diagnostic_severity}
\alias{diagnostic_from_lint}
\title{diagnostics}
\usage{
diagnostic_range(result)

diagnostic_severity(result)

diagnostic_from_lint(result)
}
\description{
Diagnose problems in files after linting.
}
\keyword{internal}
