% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{document_link_resolve}
\alias{document_link_resolve}
\title{documentLink/resolve request handler}
\usage{
document_link_resolve(self, id, params)
}
\arguments{
\item{id}{a numeric, the id of the process that started the server}

\item{params}{a DocumentLink object}
}
\description{
Handler to the \href{https://microsoft.github.io/language-server-protocol/specification#documentLink_resolve}{documentLink/resolve} \link{Request}
}
\details{
Not implemented yet.
}
\keyword{internal}
