% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{document_uri}
\alias{document_uri}
\title{A document URI}
\usage{
document_uri(uri)
}
\arguments{
\item{uri}{a character}
}
\description{
A Uniform Resource Identifier as defined by \href{https://tools.ietf.org/html/rfc3986}{RFC 3986}.
}
\keyword{internal}
