% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_package}
\alias{is_package}
\title{check if a path is a package folder}
\usage{
is_package(rootPath)
}
\arguments{
\item{rootPath}{a character representing a path}
}
\description{
check if a path is a package folder
}
\keyword{internal}
