% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageserver.R
\docType{package}
\name{languageserver-package}
\alias{languageserver}
\alias{languageserver-package}
\title{languageserver: An implementation of the Language Server Protocol for R}
\description{
An implementation of the Language Server Protocol for R
}
\details{
An implementation of the Language Server Protocol for R
}
