% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageserver.R
\docType{package}
\name{languageserver-package}
\alias{languageserver}
\alias{languageserver-package}
\title{languageserver: An Implementation of the Language Server Protocol for R}
\description{
An implementation of the Language Server Protocol for R. The Language Server protocol is used by a client to integrate features like auto complete. See <https://microsoft.github.io/language-server-protocol> for details.
}
\details{
An implementation of the Language Server Protocol for R
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/REditorSupport/languageserver}
}

}
