% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageserver.R
\docType{package}
\name{languageserver-package}
\alias{languageserver}
\alias{languageserver-package}
\title{languageserver: Language Server Protocol}
\description{
An implementation of the Language Server Protocol
    for R. The Language Server protocol is used by an editor client to
    integrate features like auto completion. See
    <https://microsoft.github.io/language-server-protocol/> for details.
}
\details{
An implementation of the Language Server Protocol for R
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/REditorSupport/languageserver/}
}

}
\author{
\strong{Maintainer}: Randy Lai \email{randy.cs.lai@gmail.com}

Other contributors:
\itemize{
  \item Kun Ren \email{mail@renkun.me} [contributor]
}

}
