% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{location}
\alias{location}
\title{a location inside a resource}
\usage{
location(uri, range)
}
\arguments{
\item{uri}{a character, the path to a file as defined by \href{https://tools.ietf.org/html/rfc3986}{RFC 3986}}

\item{range}{a \link{range}}
}
\description{
a location inside a resource
}
