% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\docType{data}
\name{Message}
\alias{Message}
\title{Base Message class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Message
}
\description{
Base Message class
}
\keyword{internal}
