% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\docType{data}
\name{Namespace}
\alias{Namespace}
\title{A class for storing package information}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
Namespace
}
\description{
A class for storing package information
}
\keyword{internal}
