% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-general.R
\name{on_initialize}
\alias{on_initialize}
\title{initialize handler}
\usage{
on_initialize(self, id, params)
}
\arguments{
\item{self}{a \link{LanguageServer} object}

\item{id}{a numeric, the id of the process that started the server}

\item{params}{a named list, the \code{initialize} Request options

@keywords internal}
}
\description{
Handler to the \href{https://microsoft.github.io/language-server-protocol/}{initialize} \link{Request}.
}
