% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-general.R
\name{on_initialize}
\alias{on_initialize}
\title{initialize handler}
\usage{
on_initialize(self, id, params)
}
\description{
Handler to the \href{https://microsoft.github.io/language-server-protocol/specification#initialize}{initialize} \link{Request}.
}
\details{
@template self
@template id
@param params a named list, the \code{initialize} Request options

@keywords internal
}
