% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-general.R
\name{on_initialized}
\alias{on_initialized}
\title{initialized handler}
\usage{
on_initialized(self, params)
}
\arguments{
\item{params}{a named list}
}
\description{
Handler to the \href{https://microsoft.github.io/language-server-protocol/specification#initialized}{initialized} \link{Notification}.
}
\keyword{internal}
