% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{position}
\alias{position}
\title{a position in a text document}
\usage{
position(line, character)
}
\arguments{
\item{line}{an integer}

\item{character}{an integer}
}
\description{
a position in a text document
}
