% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{purl}
\alias{purl}
\title{A version of knitr::purl but it doesn't delete empty lines between blocks}
\usage{
purl(path, output)
}
\description{
A version of knitr::purl but it doesn't delete empty lines between blocks
}
\keyword{internal}
