% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{range}
\alias{range}
\title{a range in a text document}
\usage{
range(start, end)
}
\arguments{
\item{start}{a \link{position}}

\item{end}{a \link{position}}
}
\description{
a range in a text document
}
