% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{range_formatting_reply}
\alias{range_formatting_reply}
\title{format a part of a document}
\usage{
range_formatting_reply(id, uri, document, range, options)
}
\arguments{
\item{id}{a character, the id of the request,}

\item{uri}{a character, the path to a file}

\item{document}{a character, the contents of the document}

\item{range}{a \link{range}, the part of the document to format}

\item{options}{a named list of options, with a \code{tabSize} parameter}
}
\description{
format a part of a document
}
