% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize_names}
\alias{sanitize_names}
\title{sanitize package objects names}
\usage{
sanitize_names(objects)
}
\description{
Remove unwanted objects, \emph{e.g.} \verb{names<-}, \verb{\%>\%}, etc.
}
\keyword{internal}
