% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{signature_reply}
\alias{signature_reply}
\title{the response to a textDocument/signatureHelp Request}
\usage{
signature_reply(id, uri, workspace, document, point)
}
\description{
If the symbol at the current position is a function, return its arguments
(as with \code{\link[base:args]{base::args()}}).
}
\keyword{internal}
