% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{style_file}
\alias{style_file}
\title{Style a file}
\usage{
style_file(path, options)
}
\description{
This functions formats a document using \code{\link[styler:style_file]{styler::style_file()}} with the
\code{\link[styler:tidyverse_style]{styler::tidyverse_style()}} style.
}
\keyword{internal}
