% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{style_text}
\alias{style_text}
\title{Edit code style}
\usage{
style_text(text, options, indentation = "")
}
\description{
This functions formats a list of text using \code{\link[styler:style_text]{styler::style_text()}} with the
\code{\link[styler:tidyverse_style]{styler::tidyverse_style()}} style.
}
\keyword{internal}
