% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_code_action}
\alias{text_document_code_action}
\title{textDocument/codeAction request handler}
\usage{
text_document_code_action(self, id, params)
}
\arguments{
\item{id}{a numeric, the id of the process that started the server}

\item{params}{a \link{code_action_params} object}
}
\description{
Handler to the \href{https://microsoft.github.io/language-server-protocol/specification#textDocument_codeAction}{textDocument/codeAction} \link{Request}
}
\details{
Not implemented yet.
}
\keyword{internal}
