% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_code_lens}
\alias{text_document_code_lens}
\title{textDocument/codeLens request handler}
\usage{
text_document_code_lens(self, id, params)
}
\arguments{
\item{id}{a numeric, the id of the process that started the server}

\item{params}{a \link{code_lens_params} object}
}
\value{
a list of CodeLens objects
}
\description{
Handler to the \href{https://microsoft.github.io/language-server-protocol/specification#textDocument_codeLens}{textDocument/codeLens} \link{Request}
}
\details{
Not implemented yet.
}
\keyword{internal}
