% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_color_presentation}
\alias{text_document_color_presentation}
\title{textDocument/colorPresentation request handler}
\usage{
text_document_color_presentation(self, id, params)
}
\arguments{
\item{self}{a \link{LanguageServer} object}

\item{id}{a numeric, the id of the process that started the server}

\item{params}{a color_presentation_params object}
}
\description{
Handler to the \href{https://microsoft.github.io/language-server-protocol/}{textDocument/colorPresentation} \link{Request}
}
\details{
Not implemented yet.
}
\keyword{internal}
