% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-textsync.R
\name{text_document_did_change}
\alias{text_document_did_change}
\title{\code{textDocument/didChange} notification handler}
\usage{
text_document_did_change(self, params)
}
\description{
Handler to the \code{textDocument/didChange} \link{Notification}.
}
\keyword{internal}
