% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-textsync.R
\name{text_document_did_close}
\alias{text_document_did_close}
\title{textDocument/didClose notification handler}
\usage{
text_document_did_close(self, params)
}
\arguments{
\item{self}{a \link{LanguageServer} object}

\item{params}{a \link{did_close_text_document_params}}
}
\description{
Handler to the \href{https://microsoft.github.io/language-server-protocol/specification#textDocument_didClose}{textDocument/didClose} \link{Notification}
}
\keyword{internal}
