% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-textsync.R
\name{text_document_did_open}
\alias{text_document_did_open}
\title{textDocument/didOpen notification handler}
\usage{
text_document_did_open(self, params)
}
\arguments{
\item{self}{a \link{LanguageServer} object}

\item{params}{a \link{did_open_text_document_params}}
}
\description{
Handler to the \href{https://microsoft.github.io/language-server-protocol/specification#textDocument_didOpen}{textDocument/didOpen} \link{Notification}
}
\keyword{internal}
