% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_document_color}
\alias{text_document_document_color}
\title{\code{textDocument/documentColor} request handler}
\usage{
text_document_document_color(self, id, params)
}
\description{
Handler to the \code{textDocument/documentColor} \link{Request}.
}
\keyword{internal}
