% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_document_highlight}
\alias{text_document_document_highlight}
\title{\code{textDocument/documentHighlight} request handler}
\usage{
text_document_document_highlight(self, id, params)
}
\description{
Handler to the \code{textDocument/documentHighlight} \link{Request}.
}
\keyword{internal}
