% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_document_link}
\alias{text_document_document_link}
\title{textDocument/documentLink request handler}
\usage{
text_document_document_link(self, id, params)
}
\arguments{
\item{id}{a numeric, the id of the process that started the server}

\item{params}{a \link{document_link_params} object}
}
\value{
a list of DocumentLink objects
}
\description{
Handler to the \href{https://microsoft.github.io/language-server-protocol/specification#textDocument_documentLink}{textDocument/documentLink} \link{Request}
}
\details{
Not implemented yet.
}
\keyword{internal}
