% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_document_symbol}
\alias{text_document_document_symbol}
\title{textDocument/documentSymbol request handler}
\usage{
text_document_document_symbol(self, id, params)
}
\arguments{
\item{id}{a numeric, the id of the process that started the server}

\item{params}{a \link{document_symbol_params} object}
}
\description{
Handler to the \href{https://microsoft.github.io/language-server-protocol/specification#textDocument_documentSymbol}{textDocument/documentSymbol} \link{Request}
}
\details{
Not implemented yet.
}
\keyword{internal}
