% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_range_formatting}
\alias{text_document_range_formatting}
\title{textDocument/rangeFormatting request handler}
\usage{
text_document_range_formatting(self, id, params)
}
\arguments{
\item{id}{a numeric, the id of the process that started the server}

\item{params}{a \link{document_range_formatting_params} object}
}
\description{
Handler to the \href{https://microsoft.github.io/language-server-protocol/specification#textDocument_rangeFormatting}{textDocument/rangeFormatting} \link{Request}
}
\keyword{internal}
