% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-langfeatures.R
\name{text_document_references}
\alias{text_document_references}
\title{textDocument/references request handler}
\usage{
text_document_references(self, id, params)
}
\arguments{
\item{self}{a \link{LanguageServer} object}

\item{id}{a numeric, the id of the process that started the server}

\item{params}{a \link{reference_params} object}
}
\description{
Handler to the \href{https://microsoft.github.io/language-server-protocol/specification#textDocument_references}{textDocument/references} \link{Request}
}
\details{
Not implemented yet.
}
\keyword{internal}
