% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-textsync.R
\name{text_document_will_save_wait_until}
\alias{text_document_will_save_wait_until}
\title{textDocument/willSaveWaitUntil notification handler}
\usage{
text_document_will_save_wait_until(self, id, params)
}
\arguments{
\item{id}{a numeric, the id of the process that started the server}

\item{params}{a \link{will_save_text_document_params}}
}
\description{
Handler to the \href{https://microsoft.github.io/language-server-protocol/specification#textDocument_willSaveWaitUntil}{textDocument/willSaveWaitUntil} \link{Request}
}
\keyword{internal}
