% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{text_edit}
\alias{text_edit}
\title{A textual edit applicable to a text document}
\usage{
text_edit(range, new_text)
}
\arguments{
\item{range}{a \link{range}, the part of the document to replace}

\item{new_text}{a character, the text to replace}
}
\description{
A textual edit applicable to a text document
}
\keyword{internal}
