% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completion.R
\name{workspace_completion}
\alias{workspace_completion}
\title{Complete any object in the workspace}
\usage{
workspace_completion(
  workspace,
  token,
  package = NULL,
  exported_only = TRUE,
  snippet_support = NULL
)
}
\description{
Complete any object in the workspace
}
\keyword{internal}
