% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-workspace.R
\name{workspace_did_change_watched_files}
\alias{workspace_did_change_watched_files}
\title{\code{workspace/didChangeWatchedFiles} notification handler}
\usage{
workspace_did_change_watched_files(self, params)
}
\description{
Handler to the \code{workspace/didChangeWatchedFiles} \link{Notification}.
}
\keyword{internal}
