% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-workspace.R
\name{workspace_did_change_workspace_folder_params}
\alias{workspace_did_change_workspace_folder_params}
\title{workspace/didChangeWorkspaceFolders notification handler}
\usage{
workspace_did_change_workspace_folder_params(self, params)
}
\arguments{
\item{params}{a did_change_workspace_folder_params object}
}
\description{
Handler to the \href{https://microsoft.github.io/language-server-protocol/specification#workspace_didChangeWorkspaceFolders}{workspace/didChangeWorkspaceFolders} \link{Notification}.
}
\details{
Not implemented yet.
}
\keyword{internal}
