% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-workspace.R
\name{workspace_execute_command}
\alias{workspace_execute_command}
\title{workspace/executeCommand request handler}
\usage{
workspace_execute_command(self, id, params)
}
\arguments{
\item{params}{an execute_command_params object}
}
\description{
Handler to the \href{https://microsoft.github.io/language-server-protocol/specification#workspace_executeCommand}{workspace/executeCommand} \link{Request}.
}
\details{
Not implemented yet.
}
\keyword{internal}
