% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-workspace.R
\name{workspace_symbol}
\alias{workspace_symbol}
\title{workspace/symbol request handler}
\usage{
workspace_symbol(self, id, params)
}
\arguments{
\item{params}{a workspace_symbol_params object}
}
\description{
Handler to the \href{https://microsoft.github.io/language-server-protocol/specification#workspace_symbol}{workspace/symbol} \link{Request}.
}
\details{
Not implemented yet.
}
\keyword{internal}
