% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{workspace_sync}
\alias{workspace_sync}
\title{Determine workspace information for a given file}
\usage{
workspace_sync(uri, lintfile = NULL, run_lintr = TRUE)
}
\arguments{
\item{uri}{the file path}

\item{lintfile}{the actual file to lint}

\item{run_lintr}{set \code{FALSE} to disable lintr diagnostics}
}
\description{
internal use only
}
