% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{workspace_sync}
\alias{workspace_sync}
\title{Determine workspace information for a given file}
\usage{
workspace_sync(uri, temp_dir = NULL, temp_file = NULL,
  run_lintr = TRUE, parse = FALSE, resolve = FALSE)
}
\arguments{
\item{uri}{the file uri}

\item{temp_file}{the file to lint, determine from \code{uri} if \code{NULL}}

\item{run_lintr}{set \code{FALSE} to disable lintr diagnostics}

\item{parse}{set \code{FALSE} to disable parsing file}
}
\description{
internal use only
}
