% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{workspace_sync}
\alias{workspace_sync}
\title{Determine workspace information for a given file}
\usage{
workspace_sync(uri, document)
}
\arguments{
\item{uri}{the file path}

\item{document}{the content of the file}
}
\description{
internal use only
}
